Run main functions
