+++ src/mirrulations_client/__main__.py
@@ -1,0 +1,9 @@
+from mirrulations_client.client import do_work
+
+
+def main():
+    do_work()
+
+
+if __name__ == '__main__':
+    main()
