+++ src/mirrulations_web/__main__.py
@@ -1,0 +1,6 @@
+def main():
+    return None
+
+
+if __name__ == '__main__':
+    main()
